Squashed commit of the following:
commit 1bf590c653086ecfb3642158c69200b5f79c247b
Author: Gogs <gogs@fake.local>
Date:   Tue Feb 05 15:08:58 2019 +0200

	Validation with custom messages from prop file completed

commit 519de046ff40896c94dfadc8cc2bd830c4f00d89
Author: Gogs <gogs@fake.local>
Date:   Tue Feb 05 12:36:41 2019 +0200

	Error messages properties file created as long with bean file which is
processing the configuration

commit eeddfda36f339f4916f4cd7d071735dea28e78f4
Author: Gogs <gogs@fake.local>
Date:   Tue Feb 05 11:48:52 2019 +0200

	Body error handling completed

commit 68a9d4b0cd673701d137c35cf52f5ed20e862d08
Author: Gogs <gogs@fake.local>
Date:   Mon Feb 04 17:58:58 2019 +0200

	field validator error in progress

commit 09802d080195ff52a526433db7997ac3ba3f0997
Author: Gogs <gogs@fake.local>
Date:   Mon Feb 04 17:58:30 2019 +0200

	The rest validation handler in progress

commit ee1b982ece1dd1af2f8218e105f9831c048346c5
Author: Gogs <gogs@fake.local>
Date:   Mon Feb 04 15:32:57 2019 +0200

	Custom error type method is added and the error handling

+++ UserRegistrationSystem/src/main/java/com/atos/boyan/Rest/UserRegistrationRestController.java
@@ -2,6 +2,8 @@ package com.atos.boyan.Rest;
 
 import java.util.List;
 
+import javax.validation.Valid;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,6 +19,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.atos.boyan.Exception.CustomErrorType;
 import com.atos.boyan.dto.UsersDTO;
 import com.atos.boyan.repository.UserJpaRepository;
 
@@ -36,11 +39,19 @@ public class UserRegistrationRestController {
 	@GetMapping("/")
 	public ResponseEntity<List<UsersDTO>> listAllUsers() {
 		List<UsersDTO> users = userJpaRepository.findAll();
+		if (users.isEmpty()) {
+			return new ResponseEntity<List<UsersDTO>>(HttpStatus.NO_CONTENT);
+		};
 		return new ResponseEntity<List<UsersDTO>>(users, HttpStatus.OK);
 	}
 
 	@PostMapping(value = "/", consumes = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<UsersDTO> createUser(@RequestBody final UsersDTO user) {
+	public ResponseEntity<UsersDTO> createUser(@Valid @RequestBody final UsersDTO user) {
+		if (userJpaRepository.findByName(user.getName()) != null) {
+			return new ResponseEntity<UsersDTO>(
+					new CustomErrorType("Unable to create user. User with name "
+							+ user + " already exist." ), HttpStatus.FOUND);
+		};
 		userJpaRepository.save(user);
 		return new ResponseEntity<UsersDTO>(user, HttpStatus.CREATED);
 	}
@@ -48,6 +59,11 @@ public class UserRegistrationRestController {
 	@GetMapping("/{id}")
 	public ResponseEntity<UsersDTO> getUserById(@PathVariable("id") final Long id) {
 		UsersDTO user = userJpaRepository.findById(id);
+		if (user == null) {
+			return new ResponseEntity<UsersDTO>(
+					new CustomErrorType("User with id "
+					+ id + " not found"), HttpStatus.NOT_FOUND);
+		}
 		return new ResponseEntity<UsersDTO>(user, HttpStatus.OK);
 	}
 
@@ -56,6 +72,12 @@ public class UserRegistrationRestController {
 			@RequestBody UsersDTO user) {
 		// fetch based on id and set it to currentUser object of type UsersDTO
 		UsersDTO currentUser = userJpaRepository.findById(id);
+		// check for wrong user id
+		if (currentUser == null) {
+			return new ResponseEntity<UsersDTO>(
+					new CustomErrorType("Unable to update! The user with id " + id
+							+ " not exist."), HttpStatus.NOT_FOUND);
+		}
 
 		// update currentUser object data with user object data
 		currentUser.setName(user.getName());
@@ -72,6 +94,12 @@ public class UserRegistrationRestController {
 
 	@DeleteMapping("/{id}")
 	public ResponseEntity<UsersDTO> deleteUser(@PathVariable("id") final Long id) {
+		UsersDTO user = userJpaRepository.findById(id);
+		if (user == null) {
+			return new ResponseEntity<UsersDTO>(
+					new CustomErrorType("Unable to delete! The user with id " + id
+							+ " not exit."), HttpStatus.NOT_FOUND);
+		};
 		userJpaRepository.delete(id);
 		return new ResponseEntity<UsersDTO>(HttpStatus.NO_CONTENT);
 	}
